+++ src/types/query.d.ts
@@ -52,6 +52,8 @@ export module query {
   export function Login(ref: ExprArg, params: ExprArg): Expr;
   export function Logout(delete_tokens: ExprArg): Expr;
   export function Identify(ref: ExprArg, password: ExprArg): Expr;
+  export function Identity(): Expr;
+  export function HasIdentity(): Expr;
 
   export function Concat(strings: ExprArg, separator?: ExprArg): Expr;
   export function Casefold(string: ExprArg, normalizer?: ExprArg): Expr;